requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cloudinary==1.8.0
coreapi==2.3.1
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.5
django-bootstrap3==9.0.0
django-ckeditor==5.3.0
django-cloudinary-storage==0.2.2
django-cors-headers==2.1.0
django-import-export==0.5.1
django-js-asset==0.1.1
django-queryset-csv==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.6.4
et-xmlfile==1.0.1
funcsigs==1.0.2
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
MySQL-python==1.2.5
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
pbr==3.1.1
Pillow==4.2.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
